+++ src/error/unsupported-extension-error.ts
@@ -1,0 +1,1 @@
+export class UnsupportedExtensionError extends Error {}