+++ src/lib/format-text/index.ts
@@ -1,4 +1,5 @@
 import { SortierOptions } from "../../config/index.js";
+import { UnsupportedExtensionError } from "../../error/unsupported-extension-error.js";
 import { getReprinterForFile } from "../../language.js";
 
 export function formatText(
@@ -9,7 +10,9 @@ export function formatText(
   const fakeFileName = `example.${fileExtension}`;
   const language = getReprinterForFile(fakeFileName);
   if (language == null) {
-    return;
+    throw new UnsupportedExtensionError(
+      `File extension ${fileExtension} is not supported`
+    );
   }
 
   const newFileContents = language.getRewrittenContents(